requirements-test.txt
pytest
pytest-asyncio
pytest-homeassistant-custom-component
aiohttp
